Add files via upload
Description: Pharmacy management system is a management system that is designed to improve accuracy and to enhance safety and efficiency in the pharmaceutical store. It is a computer based system which helps the Pharmacist to improve inventory management, cost, medical safety etc.At present, manual system is being utilized in the pharmacy. It requires the pharmacist to manually monitor each drug that is available in the pharmacy. This usually leads to mistakes as the workload of the pharmacist increases.
Technology Used:
LANGUAGE- PHP
DATABASE- MY SQL
DESIGN- HTML,CSS,JAVASCRIPT
SUPPORTS- WINDOWS ,MAC OS,LINUX